Watch icon and update our badge if it did
This works by first observing the icon to change to base64 representation (first it's an ico file), this will be the default icon.
Then compare the current icon to the default icon. If it is not the default, we assume it's a badge and we up Ferdium's badge as well.